requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.2
django-bootstrap4==23.2
django-ckeditor==6.6.1
django-js-asset==2.1.0
Pillow==10.0.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
